Address second-round review feedback
Move the shared TDS test plumbing into a neutral test_packet_support module,
delete the test-only NetworkTransport::receive, reject payload-free non-EOM
packets, and record the two production bug fixes in the changelog.

Copilot-Session: 7d033215-938c-453d-b723-a28ced51a2cb
