+++ mssql-tds/src/token/parsers/error_parser.rs
@@ -127,8 +127,8 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::io::packet_reader::tests::{TestPacketBuilder, create_network_transport_with_data};
     use crate::message::messages::PacketType;
+    use crate::test_packet_support::{TestPacketBuilder, create_network_transport_with_data};
 
     fn encode_utf16_string(s: &str) -> Vec<u8> {
         let utf16_units: Vec<u16> = s.encode_utf16().collect();
@@ -180,7 +180,7 @@ mod tests {
         // Line number
         builder.append_u32(1);
 
-        let mut packet_reader = create_network_transport_with_data(&builder.build()).await;
+        let mut packet_reader = create_network_transport_with_data(&builder.build());
 
         let parser = ErrorTokenParser::default();
         let context = ParserContext::default();
@@ -232,7 +232,7 @@ mod tests {
 
         builder.append_u32(123); // line number in proc
 
-        let mut packet_reader = create_network_transport_with_data(&builder.build()).await;
+        let mut packet_reader = create_network_transport_with_data(&builder.build());
 
         let parser = ErrorTokenParser::default();
         let context = ParserContext::default();
@@ -282,7 +282,7 @@ mod tests {
         builder.append_byte(0); // no proc name
         builder.append_u32(5);
 
-        let mut packet_reader = create_network_transport_with_data(&builder.build()).await;
+        let mut packet_reader = create_network_transport_with_data(&builder.build());
 
         let parser = ErrorTokenParser::default();
         let context = ParserContext::default();
@@ -329,7 +329,7 @@ mod tests {
         builder.append_byte(0);
         builder.append_u32(0);
 
-        let mut packet_reader = create_network_transport_with_data(&builder.build()).await;
+        let mut packet_reader = create_network_transport_with_data(&builder.build());
 
         let parser = ErrorTokenParser::default();
         let context = ParserContext::default();
